Add SubSystemController
Add a SubSystemController and edit TestControll & TestControll 2
