+++ System prototype for S&S toy Co (4915m Project)/Program.cs
@@ -12,7 +12,7 @@ namespace System_prototype_for_S_S_toy_Co__4915m_Project_
             // To customize application configuration such as set high DPI settings or default font,
             // see https://aka.ms/applicationconfiguration.
             ApplicationConfiguration.Initialize();
-            Application.Run(new Form2());
+            Application.Run(new Form1());
         }
     }
 }
